requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
black==24.3.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
ecdsa==0.18.0
fastapi==0.110.0
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.0.0
openai==1.35.3
packaging==24.0
passlib==1.7.4
pathspec==0.12.1
pgvector==0.2.5
platformdirs==4.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.36.3
tqdm==4.66.4
typing_extensions==4.10.0
urllib3==2.2.2
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wonderwords==2.2.0
